+++ src/pipecat/services/deepgram.py
@@ -32,7 +32,7 @@ class DeepgramTTSService(TTSService):
     def can_generate_metrics(self) -> bool:
         return True
 
-    async def run_tts(self, text: str) -> AsyncGenerator[Frame, None]:
+    async def run_tts(self, text: str):
         logger.debug(f"Generating TTS: [{text}]")
 
         base_url = "https://api.deepgram.com/v1/speak"
@@ -46,12 +46,12 @@ class DeepgramTTSService(TTSService):
                 if r.status != 200:
                     text = await r.text()
                     logger.error(f"Error getting audio (status: {r.status}, error: {text})")
-                    yield ErrorFrame(f"Error getting audio (status: {r.status}, error: {text})")
+                    await self.push_error(ErrorFrame(f"Error getting audio (status: {r.status}, error: {text})"))
                     return
 
                 async for data in r.content:
                     await self.stop_ttfb_metrics()
                     frame = AudioRawFrame(audio=data, sample_rate=16000, num_channels=1)
-                    yield frame
+                    await self.push_service_frame(frame)
         except Exception as e:
             logger.error(f"Deepgram exception: {e}")