+++ src/pipecat/services/moondream.py
@@ -8,8 +8,6 @@ import asyncio
 
 from PIL import Image
 
-from typing import AsyncGenerator
-
 from pipecat.frames.frames import ErrorFrame, Frame, TextFrame, VisionImageRawFrame
 from pipecat.services.ai_services import VisionService
 
@@ -69,10 +67,10 @@ class MoondreamService(VisionService):
 
         logger.debug("Loaded Moondream model")
 
-    async def run_vision(self, frame: VisionImageRawFrame) -> AsyncGenerator[Frame, None]:
+    async def run_vision(self, frame: VisionImageRawFrame):
         if not self._model:
             logger.error("Moondream model not available")
-            yield ErrorFrame("Moondream model not available")
+            await self.push_error(ErrorFrame("Moondream model not available"))
             return
 
         logger.debug(f"Analyzing image: {frame}")
@@ -88,4 +86,4 @@ class MoondreamService(VisionService):
 
         description = await asyncio.to_thread(get_image_description, frame)
 
-        yield TextFrame(text=description)
+        await self.push_service_frame(TextFrame(text=description))